requirements.txt
# FastAPI и зависимости
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# База данных
sqlalchemy[asyncio]==2.0.35
aiosqlite==0.20.0
asyncpg==0.31.0

# Безопасность
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0

# Обработка данных
pandas
openpyxl==3.1.5

# Планировщик задач
apscheduler==3.10.4

# Конфигурация
pydantic-settings==2.6.1
python-dotenv==1.0.1

# Логирование
python-json-logger==2.0.7

aiofiles
